Extract SSL configuration handling to a separate class.
